Break out decoded position logic into own file.
