+++ src/views/TrackPositionAnimation/TPADecodedPositionLogic.tsx
@@ -1,0 +1,90 @@
+import { matrix, Matrix, multiply } from "mathjs";
+import { useMemo } from "react";
+import { DecodedPositionData, DecodedPositionFrame, DecodedPositionFramePx } from "./TrackPositionAnimationTypes";
+import { computeTrackBinPixelDimensions } from "./TrackPositionAnimationView";
+
+const nullDecodedData: DecodedPositionData = {
+    type: 'DecodedPositionData',
+    xmin: 0,
+    xwidth: 0,
+    xcount: 0,
+    ymin: 0,
+    ywidth: 0,
+    ycount: 0,
+    uniqueLocations: undefined,
+    frameBounds: [],
+    values: [],
+    locations: []
+}
+
+type DecodedProbabilityLocationsMap = {
+    [linearLocation: number]: number[]
+}
+type DecodedProbabilityCorners = {
+    linearizedValue: number[]
+    nativeXUlCorner: number[]
+    nativeYUlCorner: number[]
+}
+const useUniqueDecodedUlLocations = (decodedData: DecodedPositionData | undefined): DecodedProbabilityCorners => {
+    const { locations, uniqueLocations, xcount, xwidth, xmin, ycount, ywidth, ymin } = decodedData ? decodedData : nullDecodedData
+    const mappedUlCorners = useMemo(() => {
+        const uSet = uniqueLocations ? uniqueLocations : new Set(locations).values()
+        const sortedLocations = [...uSet].sort()
+        const nativeXs = sortedLocations.map(l => l % xcount)
+        const nativeYs = sortedLocations.map(l => Math.floor(l / xcount))
+
+        const centerToULCornerMatrix = matrix([[xwidth,    0  , -xwidth/2 + xmin],
+                                               [  0   , ywidth,  ywidth/2 + ymin]])
+        const augmentedNativeCenters = matrix([nativeXs, nativeYs, new Array(nativeYs.length).fill(1)])
+        const nativeUlPoints = multiply(centerToULCornerMatrix, augmentedNativeCenters).valueOf() as number[][]
+        return {
+            linearizedValue: sortedLocations,
+            nativeXUlCorner: nativeUlPoints[0],
+            nativeYUlCorner: nativeUlPoints[1]
+        }
+    }, [locations, uniqueLocations, xcount, xwidth, xmin, ywidth, ymin])
+
+    return mappedUlCorners
+}
+
+export const useProbabilityLocationsMap = (transform: Matrix, decodedData: DecodedPositionData | undefined): DecodedProbabilityLocationsMap => {
+    const { xwidth, ywidth } = decodedData ? decodedData : nullDecodedData
+    const uniqueNativeLocations = useUniqueDecodedUlLocations(decodedData)
+    const linearPositionMap = useMemo(() => {
+        const pixelRects = computeTrackBinPixelDimensions(transform, [uniqueNativeLocations.nativeXUlCorner, uniqueNativeLocations.nativeYUlCorner], xwidth, ywidth)
+        const the_dict: DecodedProbabilityLocationsMap = {}
+        uniqueNativeLocations.linearizedValue.forEach((linearizedValue, index) => {
+            the_dict[linearizedValue] = pixelRects[index]
+        })
+        return the_dict
+    }, [transform, uniqueNativeLocations, xwidth, ywidth])
+    return linearPositionMap
+}
+
+export const useProbabilityFrames = (decodedData: DecodedPositionData | undefined): DecodedPositionFrame[] => {
+    const { frameBounds, values, locations } = decodedData ? decodedData : nullDecodedData
+
+    const frames = useMemo(() => {
+        let frameStart = 0
+        return frameBounds.map((nObservations) => {
+            const valueSlice = values.slice(frameStart, frameStart + nObservations)
+            const locationSlice = locations.slice(frameStart, frameStart + nObservations)
+            frameStart += nObservations
+            return { linearLocations: locationSlice, values: valueSlice }
+        })
+    }, [locations, values, frameBounds])
+
+    return frames
+}
+
+export const getDecodedPositionFramePx = (linearFrame: DecodedPositionFrame | undefined, decodedLocationsMap: DecodedProbabilityLocationsMap): DecodedPositionFramePx | undefined => {
+    const pixelLocations = linearFrame ? linearFrame.linearLocations.map((l) => decodedLocationsMap[l]) : []
+    // console.log(`frame # ${animationState.currentFrameIndex}\nLinear pairs: ${linearFrame?.linearLocations} vs ${linearFrame?.values}`)
+    const finalFrame = linearFrame
+        ? {
+            locationRectsPx: pixelLocations,
+            values: linearFrame.values
+        }
+        : undefined
+    return finalFrame
+}
